添加 Vector3 和 Vector4 类，更新相关测试文件，修复关于 Linux 系统的编译问题

+++ LikesProgram/src/LikesProgram/String.cpp
@@ -885,7 +885,7 @@ namespace LikesProgram {
     String String::FromFloat(double value, size_t precision) {
         std::wostringstream woss;
         if (precision > 0) {
-            woss << std::fixed << std::setprecision(precision);
+            woss << std::fixed << std::setprecision((int)precision);
         }
         woss << value;
         return String(woss.str());
